texstudio: search recursive for files
When using ls we list only files with the corresonding extension in the
current directory. Assuming the following structure:

- dir1
- example.tex
- dir2
	- example2.tex

The old verison would only list example.tex for completion. But maybe you
want to edit file example2.tex in the subdirectory.
The find command search recursivly for the files with the coressonding
extension thus all matching files in subdirectories are considered as well.